backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pymongo>=4.6.1
motor>=3.3.2
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0


